requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
python-dateutil